requirements.txt
pandas
numpy
matplotlib
seaborn
plotly
scikit-learn
streamlit_extras
mysql-connector-python
